databricks-apps/sample-get-uc-service-creds/requirements.txt
streamlit==1.38.0
pandas>=2.0.0
boto3>=1.26.0
databricks-sdk>=0.63.0
python-dotenv>=1.0.0
